+++ app/pages/SiloUsersAndGroupsGroupsTab.tsx
@@ -5,10 +5,8 @@
  *
  * Copyright Oxide Computer Company
  */
-import { useQuery } from '@tanstack/react-query'
 import { createColumnHelper } from '@tanstack/react-table'
 import { useCallback, useMemo, useState } from 'react'
-import * as R from 'remeda'
 
 import {
   api,
@@ -16,18 +14,15 @@ import {
   getListQFn,
   q,
   queryClient,
-  roleOrder,
+  rolesByIdFromPolicy,
   useApiMutation,
   usePrefetchedQuery,
   type Group,
-  type Policy,
-  type RoleKey,
-  type User,
 } from '@oxide/api'
-import { PersonGroup16Icon, PersonGroup24Icon } from '@oxide/design-system/icons/react'
+import { PersonGroup24Icon } from '@oxide/design-system/icons/react'
 import { Badge } from '@oxide/design-system/ui'
 
-import { ReadOnlySideModalForm } from '~/components/form/ReadOnlySideModalForm'
+import { GroupMembersSideModal } from '~/components/access/GroupMembersSideModal'
 import { HL } from '~/components/HL'
 import { SiloAccessEditUserSideModal } from '~/forms/silo-access'
 import { titleCrumb } from '~/hooks/use-crumbs'
@@ -36,15 +31,11 @@ import { addToast } from '~/stores/toast'
 import { EmptyCell } from '~/table/cells/EmptyCell'
 import { ButtonCell } from '~/table/cells/LinkCell'
 import { MemberCountCell } from '~/table/cells/MemberCountCell'
-import { RowActions, useColsWithActions, type MenuAction } from '~/table/columns/action-col'
+import { useColsWithActions, type MenuAction } from '~/table/columns/action-col'
 import { Columns } from '~/table/columns/common'
 import { useQueryTable } from '~/table/QueryTable'
 import { EmptyMessage } from '~/ui/lib/EmptyMessage'
-import { PropertiesTable } from '~/ui/lib/PropertiesTable'
-import { ResourceLabel } from '~/ui/lib/SideModal'
-import { Table } from '~/ui/lib/Table'
 import { roleColor } from '~/util/access'
-import { ALL_ISH } from '~/util/consts'
 
 const policyView = q(api.policyView, {})
 const groupList = getListQFn(api.groupList, {})
@@ -69,116 +60,6 @@ const GroupEmptyState = () => (
   />
 )
 
-type GroupMembersSideModalProps = {
-  group: Group
-  onDismiss: () => void
-  siloPolicy: Policy
-}
-
-type SiloGroupRoleEntry = {
-  scope: 'silo'
-  roleName: RoleKey
-  source: { type: 'direct' }
-}
-
-function GroupMembersSideModal({
-  group,
-  onDismiss,
-  siloPolicy,
-}: GroupMembersSideModalProps) {
-  const { data } = useQuery(q(api.userList, { query: { group: group.id, limit: ALL_ISH } }))
-  const members = data?.items ?? []
-
-  const roleEntries: SiloGroupRoleEntry[] = []
-  const directAssignment = siloPolicy.roleAssignments.find(
-    (ra) => ra.identityId === group.id
-  )
-  if (directAssignment) {
-    roleEntries.push({
-      scope: 'silo',
-      roleName: directAssignment.roleName,
-      source: { type: 'direct' },
-    })
-  }
-  const sortedRoleEntries = R.sortBy(roleEntries, (e) => roleOrder[e.roleName])
-
-  return (
-    <ReadOnlySideModalForm
-      title="Group"
-      subtitle={
-        <ResourceLabel>
-          <PersonGroup16Icon /> {group.displayName}
-        </ResourceLabel>
-      }
-      onDismiss={onDismiss}
-      animate
-    >
-      <PropertiesTable>
-        <PropertiesTable.IdRow id={group.id} />
-        <PropertiesTable.DateRow label="Created" date={group.timeCreated} />
-      </PropertiesTable>
-      <div className="mt-6">
-        <table className="ox-table text-sans-md w-full border-separate">
-          <Table.Header>
-            <Table.HeaderRow>
-              <Table.HeadCell>Role</Table.HeadCell>
-              <Table.HeadCell>Source</Table.HeadCell>
-            </Table.HeaderRow>
-          </Table.Header>
-          <Table.Body>
-            {sortedRoleEntries.length === 0 ? (
-              <Table.Row>
-                <Table.Cell colSpan={2} className="text-secondary">
-                  No roles assigned
-                </Table.Cell>
-              </Table.Row>
-            ) : (
-              sortedRoleEntries.map(({ scope, roleName }, i) => (
-                <Table.Row key={i}>
-                  <Table.Cell>
-                    <Badge color={roleColor[roleName]}>
-                      {scope}.{roleName}
-                    </Badge>
-                  </Table.Cell>
-                  <Table.Cell>Assigned</Table.Cell>
-                </Table.Row>
-              ))
-            )}
-          </Table.Body>
-        </table>
-      </div>
-      <div className="mt-6">
-        {members.length === 0 ? (
-          <EmptyMessage
-            icon={<PersonGroup24Icon />}
-            title="No members"
-            body="This group has no members"
-          />
-        ) : (
-          <table className="ox-table text-sans-md w-full border-separate">
-            <Table.Header>
-              <Table.HeaderRow>
-                <Table.HeadCell>Members</Table.HeadCell>
-                <Table.HeadCell />
-              </Table.HeaderRow>
-            </Table.Header>
-            <Table.Body>
-              {members.map((member: User) => (
-                <Table.Row key={member.id}>
-                  <Table.Cell>{member.displayName}</Table.Cell>
-                  <Table.Cell className="action-col w-10 *:p-0">
-                    <RowActions id={member.id} />
-                  </Table.Cell>
-                </Table.Row>
-              ))}
-            </Table.Body>
-          </table>
-        )}
-      </div>
-    </ReadOnlySideModalForm>
-  )
-}
-
 export default function SiloUsersAndGroupsGroupsTab() {
   const [selectedGroup, setSelectedGroup] = useState<Group | null>(null)
   const [editingGroup, setEditingGroup] = useState<Group | null>(null)
@@ -192,10 +73,7 @@ export default function SiloUsersAndGroupsGroupsTab() {
     },
   })
 
-  const siloRoleById = useMemo(
-    () => new Map(siloPolicy.roleAssignments.map((a) => [a.identityId, a.roleName])),
-    [siloPolicy]
-  )
+  const siloRoleById = useMemo(() => rolesByIdFromPolicy(siloPolicy), [siloPolicy])
 
   const siloRoleCol = useMemo(
     () =>
@@ -268,7 +146,7 @@ export default function SiloUsersAndGroupsGroupsTab() {
         <GroupMembersSideModal
           group={selectedGroup}
           onDismiss={() => setSelectedGroup(null)}
-          siloPolicy={siloPolicy}
+          scopedPolicies={[{ scope: 'silo', policy: siloPolicy, sourceLabel: 'Assigned' }]}
         />
       )}
       {editingGroup && (
