+++ app/pages/project/access/ProjectUsersAndGroupsGroupsTab.tsx
@@ -5,7 +5,6 @@
  *
  * Copyright Oxide Computer Company
  */
-import { useQuery } from '@tanstack/react-query'
 import { createColumnHelper } from '@tanstack/react-table'
 import { useCallback, useMemo, useState } from 'react'
 import type { LoaderFunctionArgs } from 'react-router'
@@ -18,17 +17,15 @@ import {
   q,
   queryClient,
   roleOrder,
+  rolesByIdFromPolicy,
   useApiMutation,
   usePrefetchedQuery,
   type Group,
-  type Policy,
-  type RoleKey,
-  type User,
 } from '@oxide/api'
-import { PersonGroup16Icon, PersonGroup24Icon } from '@oxide/design-system/icons/react'
+import { PersonGroup24Icon } from '@oxide/design-system/icons/react'
 import { Badge } from '@oxide/design-system/ui'
 
-import { ReadOnlySideModalForm } from '~/components/form/ReadOnlySideModalForm'
+import { GroupMembersSideModal } from '~/components/access/GroupMembersSideModal'
 import { HL } from '~/components/HL'
 import { ListPlusCell } from '~/components/ListPlusCell'
 import { ProjectAccessEditUserSideModal } from '~/forms/project-access'
@@ -39,16 +36,12 @@ import { addToast } from '~/stores/toast'
 import { EmptyCell } from '~/table/cells/EmptyCell'
 import { ButtonCell } from '~/table/cells/LinkCell'
 import { MemberCountCell } from '~/table/cells/MemberCountCell'
-import { RowActions, useColsWithActions, type MenuAction } from '~/table/columns/action-col'
+import { useColsWithActions, type MenuAction } from '~/table/columns/action-col'
 import { Columns } from '~/table/columns/common'
 import { useQueryTable } from '~/table/QueryTable'
 import { EmptyMessage } from '~/ui/lib/EmptyMessage'
-import { PropertiesTable } from '~/ui/lib/PropertiesTable'
-import { ResourceLabel } from '~/ui/lib/SideModal'
-import { Table } from '~/ui/lib/Table'
 import { TipIcon } from '~/ui/lib/TipIcon'
 import { roleColor } from '~/util/access'
-import { ALL_ISH } from '~/util/consts'
 import type * as PP from '~/util/path-params'
 
 const policyView = q(api.policyView, {})
@@ -78,133 +71,6 @@ const GroupEmptyState = () => (
   />
 )
 
-type GroupMembersSideModalProps = {
-  group: Group
-  onDismiss: () => void
-  siloPolicy: Policy
-  projectPolicy: Policy
-}
-
-type ProjectGroupRoleEntry = {
-  scope: 'silo' | 'project'
-  roleName: RoleKey
-  source: { type: 'direct' } | { type: 'silo' }
-}
-
-function GroupMembersSideModal({
-  group,
-  onDismiss,
-  siloPolicy,
-  projectPolicy,
-}: GroupMembersSideModalProps) {
-  const { data } = useQuery(q(api.userList, { query: { group: group.id, limit: ALL_ISH } }))
-  const members = data?.items ?? []
-
-  const roleEntries: ProjectGroupRoleEntry[] = []
-
-  const directProjectAssignment = projectPolicy.roleAssignments.find(
-    (ra) => ra.identityId === group.id
-  )
-  if (directProjectAssignment) {
-    roleEntries.push({
-      scope: 'project',
-      roleName: directProjectAssignment.roleName,
-      source: { type: 'direct' },
-    })
-  }
-
-  const directSiloAssignment = siloPolicy.roleAssignments.find(
-    (ra) => ra.identityId === group.id
-  )
-  if (directSiloAssignment) {
-    roleEntries.push({
-      scope: 'silo',
-      roleName: directSiloAssignment.roleName,
-      source: { type: 'silo' },
-    })
-  }
-
-  const sortedRoleEntries = R.sortBy(roleEntries, (e) => roleOrder[e.roleName])
-
-  return (
-    <ReadOnlySideModalForm
-      title="Group"
-      subtitle={
-        <ResourceLabel>
-          <PersonGroup16Icon /> {group.displayName}
-        </ResourceLabel>
-      }
-      onDismiss={onDismiss}
-      animate
-    >
-      <PropertiesTable>
-        <PropertiesTable.IdRow id={group.id} />
-        <PropertiesTable.DateRow label="Created" date={group.timeCreated} />
-      </PropertiesTable>
-      <div className="mt-6">
-        <table className="ox-table text-sans-md w-full border-separate">
-          <Table.Header>
-            <Table.HeaderRow>
-              <Table.HeadCell>Role</Table.HeadCell>
-              <Table.HeadCell>Source</Table.HeadCell>
-            </Table.HeaderRow>
-          </Table.Header>
-          <Table.Body>
-            {sortedRoleEntries.length === 0 ? (
-              <Table.Row>
-                <Table.Cell colSpan={2} className="text-secondary">
-                  No roles assigned
-                </Table.Cell>
-              </Table.Row>
-            ) : (
-              sortedRoleEntries.map(({ scope, roleName, source }, i) => (
-                <Table.Row key={i}>
-                  <Table.Cell>
-                    <Badge color={roleColor[roleName]}>
-                      {scope}.{roleName}
-                    </Badge>
-                  </Table.Cell>
-                  <Table.Cell>
-                    {source.type === 'direct' ? 'Assigned' : 'Inherited from silo'}
-                  </Table.Cell>
-                </Table.Row>
-              ))
-            )}
-          </Table.Body>
-        </table>
-      </div>
-      <div className="mt-6">
-        {members.length === 0 ? (
-          <EmptyMessage
-            icon={<PersonGroup24Icon />}
-            title="No members"
-            body="This group has no members"
-          />
-        ) : (
-          <table className="ox-table text-sans-md w-full border-separate">
-            <Table.Header>
-              <Table.HeaderRow>
-                <Table.HeadCell>Members</Table.HeadCell>
-                <Table.HeadCell />
-              </Table.HeaderRow>
-            </Table.Header>
-            <Table.Body>
-              {members.map((member: User) => (
-                <Table.Row key={member.id}>
-                  <Table.Cell>{member.displayName}</Table.Cell>
-                  <Table.Cell className="action-col w-10 *:p-0">
-                    <RowActions id={member.id} />
-                  </Table.Cell>
-                </Table.Row>
-              ))}
-            </Table.Body>
-          </table>
-        )}
-      </div>
-    </ReadOnlySideModalForm>
-  )
-}
-
 export default function ProjectUsersAndGroupsGroupsTab() {
   const [selectedGroup, setSelectedGroup] = useState<Group | null>(null)
   const [editingGroup, setEditingGroup] = useState<Group | null>(null)
@@ -221,14 +87,8 @@ export default function ProjectUsersAndGroupsGroupsTab() {
     },
   })
 
-  const siloRoleById = useMemo(
-    () => new Map(siloPolicy.roleAssignments.map((a) => [a.identityId, a.roleName])),
-    [siloPolicy]
-  )
-  const projectRoleById = useMemo(
-    () => new Map(projectPolicy.roleAssignments.map((a) => [a.identityId, a.roleName])),
-    [projectPolicy]
-  )
+  const siloRoleById = useMemo(() => rolesByIdFromPolicy(siloPolicy), [siloPolicy])
+  const projectRoleById = useMemo(() => rolesByIdFromPolicy(projectPolicy), [projectPolicy])
 
   const rolesCol = useMemo(
     () =>
@@ -331,8 +191,10 @@ export default function ProjectUsersAndGroupsGroupsTab() {
         <GroupMembersSideModal
           group={selectedGroup}
           onDismiss={() => setSelectedGroup(null)}
-          siloPolicy={siloPolicy}
-          projectPolicy={projectPolicy}
+          scopedPolicies={[
+            { scope: 'project', policy: projectPolicy, sourceLabel: 'Assigned' },
+            { scope: 'silo', policy: siloPolicy, sourceLabel: 'Inherited from silo' },
+          ]}
         />
       )}
       {editingGroup && (
